refactor: move para camada de aplicação as abstrações relacionadas à Storage
