+++ application/src/main/java/com/arqivame/storage/application/service/storage/FileAssemblerService.java
@@ -1,4 +1,4 @@
-package com.arqivame.storage.domain.file.service;
+package com.arqivame.storage.application.service.storage;
 
 import java.util.Objects;
 import java.util.Set;
