+++ application/src/main/java/com/arqivame/storage/application/service/storage/StorageKey.java
@@ -1,4 +1,4 @@
-package com.arqivame.storage.domain.file.service;
+package com.arqivame.storage.application.service.storage;
 
 import java.util.Arrays;
 import java.util.Objects;
