+++ application/src/main/java/com/arqivame/storage/application/service/storage/StorageReader.java
@@ -1,0 +1,14 @@
+package com.arqivame.storage.application.service.storage;
+
+import java.io.InputStream;
+
+@FunctionalInterface
+public interface StorageReader {
+
+    InputStream read(
+            StorageKey key,
+            Long offsetInBytes,
+            Long sizeInBytes,
+            Long bytesPerSecondsReadRate);
+
+}
